feat: add function to create/update .nvmrc with Node.js version
This commit introduces a new feature in the npmjs.el Emacs package
that allows users to create or update the .nvmrc file with a specified
Node.js version. Two new functions have been added: `npmjs--write-nvm-file`
which is a helper function to perform the file write operation, and
`npmjs-write-nvm-file` which is an interactive function that prompts
the user to select a Node.js version and then calls the helper to
write it to the .nvmrc file. Additionally, a new transient menu entry
has been added to facilitate this operation from the package's UI.